# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handler.SwaggerError:
    properties:
      code:
        example: VALIDATION_FAILED
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        example: Validation failed
        type: string
      requestId:
        example: req-abc123
        type: string
      timestamp:
        example: "2024-02-01T10:30:00Z"
        type: string
    type: object
  internal_handler.SwaggerErrorResponse:
    description: Standard error response format
    properties:
      error:
        $ref: '#/definitions/internal_handler.SwaggerError'
    type: object
...
  internal_handler.SwaggerPantryExpiringResponse:
    description: List of expiring pantry items
    properties:
      count:
        example: 5
        type: integer
      days:
        example: 7
        type: integer
      items:
        items:
          $ref: '#/definitions/internal_handler.SwaggerPantryItem'
        type: array
    type: object
  internal_handler.SwaggerPantryItem:
    description: Pantry inventory item
    properties:
      category:
        enum:
        - dairy
        - produce
        - proteins
        - bakery
        - pantry
        - spices
        - condiments
        - beverages
        - snacks
        - frozen
        - household
        - other
        example: condiments
        type: string
      createdAt:
        example: "2024-02-01T10:30:00Z"
        type: string
      expirationDate:
        example: "2024-06-01T00:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Olive Oil
        type: string
      quantity:
        example: 500
        type: number
      syncVersion:
        example: 1
        type: integer
      unit:
        example: ml
        type: string
      updatedAt:
        example: "2024-02-01T10:30:00Z"
        type: string
      userId:
        example: 550e8400-e29b-41d4-a716-446655440099
        type: string
    type: object
  internal_handler.SwaggerPantryItemInput:
    description: Pantry item input for create/update
    properties:
      category:
        enum:
        - dairy
        - produce
        - proteins
        - bakery
        - pantry
        - spices
        - condiments
        - beverages
        - snacks
        - frozen
        - household
        - other
        example: condiments
        type: string
      expirationDate:
        example: "2024-06-01T00:00:00Z"
        type: string
      name:
        example: Olive Oil
        type: string
      quantity:
        example: 500
        type: number
      unit:
        example: ml
        type: string
    required:
    - category
    - name
    type: object
  internal_handler.SwaggerPantryListResponse:
    description: Paginated list of pantry items
    properties:
      count:
        example: 20
        type: integer
      items:
        items:
          $ref: '#/definitions/internal_handler.SwaggerPantryItem'
        type: array
      limit:
        example: 100
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 50
        type: integer
    type: object
  internal_handler.SwaggerPantryScanRequest:
    description: AI pantry scan request
    properties:
      autoAdd:
        example: false
        type: boolean
      imageBase64:
        example: /9j/4AAQSkZJRgABAQ...
        type: string
      mimeType:
        example: image/jpeg
        type: string
    required:
    - imageBase64
    type: object
...
  internal_handler.SwaggerScanResponse:
    description: AI pantry scan results
    properties:
      addedCount:
        example: 3
        type: integer
      addedIds:
        example:
        - id1
        - id2
        - id3
        items:
          type: string
        type: array
      confidence:
        example: 0.89
        type: number
      items:
        items:
          $ref: '#/definitions/internal_handler.SwaggerScannedItem'
        type: array
      notes:
        example: Detected items from refrigerator shelf
        type: string
    type: object
  internal_handler.SwaggerScannedItem:
    description: AI-detected pantry item from scan
    properties:
      added:
        example: true
        type: boolean
      addedId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      category:
        example: dairy
        type: string
      confidence:
        example: 0.95
        type: number
      expirationDate:
        example: "2024-02-15"
        type: string
      name:
        example: Milk
        type: string
      quantity:
        example: 1
        type: number
      unit:
        example: gallon
        type: string
    type: object
...
host: api.dishflow.app
info:
  contact:
    email: support@dishflow.app
    name: DishFlow Support
  description: Recipe management API with AI-powered extraction, pantry tracking,
    and shopping lists
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://dishflow.app/terms
  title: DishFlow API
  version: "1.0"
paths:
...
  /pantry:
    get:
      description: Get paginated list of user's pantry items with optional category
        filter
      parameters:
      - description: Filter by category
        enum:
        - dairy
        - produce
        - proteins
        - bakery
        - pantry
        - spices
        - condiments
        - beverages
        - snacks
        - frozen
        - household
        - other
        in: query
        name: category
        type: string
      - default: 100
        description: Items per page (max 200)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of pantry items
          schema:
            $ref: '#/definitions/internal_handler.SwaggerPantryListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List pantry items
      tags:
      - Pantry
    post:
      consumes:
      - application/json
      description: Add a new item to the user's pantry inventory
      parameters:
      - description: Pantry item data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SwaggerPantryItemInput'
      produces:
      - application/json
      responses:
        "201":
          description: Pantry item created
          schema:
            $ref: '#/definitions/internal_handler.SwaggerPantryItem'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a pantry item
      tags:
      - Pantry
  /pantry/{id}:
    delete:
      description: Remove an item from the user's pantry
      parameters:
      - description: Pantry item UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Pantry item deleted
        "400":
          description: Invalid item ID
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "404":
          description: Pantry item not found
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a pantry item
      tags:
      - Pantry
    get:
      description: Get a single pantry item's details
      parameters:
      - description: Pantry item UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Pantry item details
          schema:
            $ref: '#/definitions/internal_handler.SwaggerPantryItem'
        "400":
          description: Invalid item ID
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "404":
          description: Pantry item not found
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get pantry item by ID
      tags:
      - Pantry
    put:
      consumes:
      - application/json
      description: Update an existing pantry item's details
      parameters:
      - description: Pantry item UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated pantry item data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SwaggerPantryItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: Pantry item updated
          schema:
            $ref: '#/definitions/internal_handler.SwaggerPantryItem'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "404":
          description: Pantry item not found
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a pantry item
      tags:
      - Pantry
  /pantry/expiring:
    get:
      description: Get list of pantry items expiring within specified days
      parameters:
      - default: 7
        description: Days until expiration (max 365)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of expiring items
          schema:
            $ref: '#/definitions/internal_handler.SwaggerPantryExpiringResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get expiring pantry items
      tags:
      - Pantry
  /pantry/scan:
    post:
      consumes:
      - multipart/form-data
      - application/json
      description: Scan image to detect and optionally add pantry items using AI
      parameters:
      - description: Image file (multipart)
        in: formData
        name: image
        type: file
      - default: false
        description: Auto-add detected items
        in: formData
        name: autoAdd
        type: boolean
      - description: JSON with base64 image
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_handler.SwaggerPantryScanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Scan results
          schema:
            $ref: '#/definitions/internal_handler.SwaggerScanResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "422":
          description: Scan failed
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: AI-powered pantry scan
      tags:
      - Pantry
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {your_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...